implement: Add new ProcessManager class to replace the old memory management implementation
